requirements.txt
opencv-python
numpy
torch
Pillow